+++ packages/input/src/Component.tsx
@@ -202,7 +202,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
 
         const [focusVisible] = useFocus(inputRef, 'keyboard');
 
-        const [focused, setFocused] = useState(false);
+        const [focused, setFocused] = useState(restProps.autoFocus);
         const [stateValue, setStateValue] = useState(defaultValue || '');
 
         const filled = Boolean(uncontrolled ? stateValue : value);
@@ -333,6 +333,7 @@ export const Input = React.forwardRef<HTMLInputElement, InputProps>(
                     className={cn(
                         styles.input,
                         {
+                            [styles.error]: error,
                             [styles.hasLabel]: label,
                         },
                         inputClassName,