+++ packages/radio/src/Component.test.tsx
@@ -126,5 +126,29 @@ describe('Radio', () => {
 
             expect(cb).not.toBeCalled();
         });
+
+        test('should not call `onChange` prop if inactive', () => {
+            const cb = jest.fn();
+
+            const { container } = render(<Radio onChange={cb} inactive={true} />);
+
+            if (container.firstElementChild) {
+                fireEvent.click(container.firstElementChild);
+            }
+
+            expect(cb).not.toBeCalled();
+        });
+
+        test('should not call `onChange` prop if inactive and checked', () => {
+            const cb = jest.fn();
+
+            const { container } = render(<Radio onChange={cb} checked={true} inactive={true} />);
+
+            if (container.firstElementChild) {
+                fireEvent.click(container.firstElementChild);
+            }
+
+            expect(cb).not.toBeCalled();
+        });
     });
 });