+++ packages/attach/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@3.0.2...@alfalab/core-components-attach@3.1.0) (2021-07-23)
+
+
+### Features
+
+* **attach:** simulate controlled state ([#760](https://github.com/alfa-laboratory/core-components/issues/760)) ([5d312bf](https://github.com/alfa-laboratory/core-components/commit/5d312bf632cc716bfc7ecb7b487c4af812838aa9)), closes [#743](https://github.com/alfa-laboratory/core-components/issues/743)
+
+
+
+
+
+## [3.0.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@3.0.1...@alfalab/core-components-attach@3.0.2) (2021-07-19)
+
+**Note:** Version bump only for package @alfalab/core-components-attach
+
+
+
+
+
 ## [3.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@3.0.0...@alfalab/core-components-attach@3.0.1) (2021-07-09)
 
 **Note:** Version bump only for package @alfalab/core-components-attach