+++ packages/confirmation/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.1.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-confirmation@8.0.2...@alfalab/core-components-confirmation@8.1.0) (2021-07-23)
+
+
+### Features
+
+* **confirmation:** fix timer, some loader refactoring ([#754](https://github.com/alfa-laboratory/core-components/issues/754)) ([187bb0e](https://github.com/alfa-laboratory/core-components/commit/187bb0e72a75c62b117a49917976903257d90bb7))
+
+
+
+
+
+## [8.0.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-confirmation@8.0.1...@alfalab/core-components-confirmation@8.0.2) (2021-07-19)
+
+**Note:** Version bump only for package @alfalab/core-components-confirmation
+
+
+
+
+
 ## [8.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-confirmation@8.0.0...@alfalab/core-components-confirmation@8.0.1) (2021-07-09)
 
 **Note:** Version bump only for package @alfalab/core-components-confirmation