+++ packages/input-autocomplete/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-input-autocomplete@7.0.3...@alfalab/core-components-input-autocomplete@7.0.4) (2021-08-03)
+
+**Note:** Version bump only for package @alfalab/core-components-input-autocomplete
+
+
+
+
+
+## [7.0.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-input-autocomplete@7.0.2...@alfalab/core-components-input-autocomplete@7.0.3) (2021-07-23)
+
+**Note:** Version bump only for package @alfalab/core-components-input-autocomplete
+
+
+
+
+
+## [7.0.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-input-autocomplete@7.0.1...@alfalab/core-components-input-autocomplete@7.0.2) (2021-07-19)
+
+**Note:** Version bump only for package @alfalab/core-components-input-autocomplete
+
+
+
+
+
 ## [7.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-input-autocomplete@7.0.0...@alfalab/core-components-input-autocomplete@7.0.1) (2021-07-09)
 
 **Note:** Version bump only for package @alfalab/core-components-input-autocomplete