+++ packages/picker-button/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-picker-button@6.1.2...@alfalab/core-components-picker-button@6.1.3) (2021-08-03)
+
+**Note:** Version bump only for package @alfalab/core-components-picker-button
+
+
+
+
+
+## [6.1.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-picker-button@6.1.1...@alfalab/core-components-picker-button@6.1.2) (2021-07-23)
+
+**Note:** Version bump only for package @alfalab/core-components-picker-button
+
+
+
+
+
+## [6.1.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-picker-button@6.1.0...@alfalab/core-components-picker-button@6.1.1) (2021-07-19)
+
+**Note:** Version bump only for package @alfalab/core-components-picker-button
+
+
+
+
+
 # [6.1.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-picker-button@6.0.0...@alfalab/core-components-picker-button@6.1.0) (2021-07-09)